Refactor plotting into astroq/plot package with PlotData context.
Replace StarPlotter/process_stars with build_plot_data and PlotSelection so plot functions share one pandas-backed context, split the monolithic plot module by domain, and add snapshot tests to guard figure output.
