+++ tests/baseline/plot/birdseye.sha256
@@ -1,0 +1,1 @@
+bc5a6425ad256f8e22b1745feb78625fcda80ad922149b589517d502fcc4bbb9
