+++ tests/baseline/plot/cof_time.sha256
@@ -1,0 +1,1 @@
+e42f7a49bade35a25075920efd4cb7823f6cdbf5a32819c36b964320251c56ff
