+++ tests/baseline/plot/cof_visits.sha256
@@ -1,0 +1,1 @@
+f43c67ffa28e55838625f1cca29a88f071e9f45e1849b5015ceaeeacfebf202c
