+++ tests/baseline/plot/completion_hist.sha256
@@ -1,0 +1,1 @@
+deb6f72a789bda3fd2d988b0b1fc4cfea06e99dda4e6820e31ff1eb0372c8697
