+++ tests/baseline/plot/completion_scatter.sha256
@@ -1,0 +1,1 @@
+ce35e5c1d28d4f368577095860b5e566956e7367c1e94a3d4c60c655bd9d0604
