+++ tests/baseline/plot/football.sha256
@@ -1,0 +1,1 @@
+a18fde724293cc5c0e3da4455d6be313be498cc9afbe01b7545cbbd55af7bef8
