+++ tests/baseline/plot/tau_inter.sha256
@@ -1,0 +1,1 @@
+3709c7db7508059e9a193a5501b30bdc5b695f03bfefe206f6bf3973770ec90a
