+++ tests/baseline/plot/timebar.sha256
@@ -1,0 +1,1 @@
+455756f15c9f3c426a38e50e69cb424d933c7ef3103ba542f32062197b96dec2
